Network.AWS.Batch

Description

AWS Batch enables you to run batch computing workloads on the AWS Cloud. Batch computing is a common way for developers, scientists, and engineers to access large amounts of compute resources, and AWS Batch removes the undifferentiated heavy lifting of configuring and managing the required infrastructure. AWS Batch will be familiar to users of traditional batch computing software. This service can efficiently provision resources in response to jobs submitted in order to eliminate capacity constraints, reduce compute costs, and deliver results quickly.

As a fully managed service, AWS Batch enables developers, scientists, and engineers to run batch computing workloads of any scale. AWS Batch automatically provisions compute resources and optimizes the workload distribution based on the quantity and scale of the workloads. With AWS Batch, there is no need to install or manage batch computing software, which allows you to focus on analyzing results and solving problems. AWS Batch reduces operational complexities, saves time, and reduces costs, which makes it easy for developers, scientists, and engineers to run their batch jobs in the AWS Cloud.

Service Configuration

batch :: Service

API version 2016-08-10 of the Amazon Batch SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Batch.

ServerException

_ServerException :: AsError a => Getting (First ServiceError) a ServiceError

These errors are usually caused by a server issue.

ClientException

_ClientException :: AsError a => Getting (First ServiceError) a ServiceError

These errors are usually caused by a client action, such as using an action or resource on behalf of a user that doesn't have permission to use the action or resource, or specifying an identifier that is not valid.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CreateComputeEnvironment

module Network.AWS.Batch.CreateComputeEnvironment

RegisterJobDefinition

module Network.AWS.Batch.RegisterJobDefinition

SubmitJob

module Network.AWS.Batch.SubmitJob

ListJobs

module Network.AWS.Batch.ListJobs

TerminateJob

module Network.AWS.Batch.TerminateJob

DescribeJobs

module Network.AWS.Batch.DescribeJobs

DeleteComputeEnvironment

module Network.AWS.Batch.DeleteComputeEnvironment

UpdateComputeEnvironment

module Network.AWS.Batch.UpdateComputeEnvironment

DescribeJobDefinitions

module Network.AWS.Batch.DescribeJobDefinitions

UpdateJobQueue

module Network.AWS.Batch.UpdateJobQueue

DeleteJobQueue

module Network.AWS.Batch.DeleteJobQueue

CreateJobQueue

module Network.AWS.Batch.CreateJobQueue

DeregisterJobDefinition

module Network.AWS.Batch.DeregisterJobDefinition

DescribeJobQueues

module Network.AWS.Batch.DescribeJobQueues

DescribeComputeEnvironments

module Network.AWS.Batch.DescribeComputeEnvironments

CancelJob

module Network.AWS.Batch.CancelJob

Types

CEState

data CEState

Constructors

Disabled

Enabled

CEStatus

data CEStatus

Constructors

CESCreating

CESDeleted

CESDeleting

CESInvalid

CESUpdating

CESValid

CEType

data CEType

Constructors

Managed

Unmanaged

CRType

data CRType

Constructors

EC2

Spot

JQState

data JQState

Constructors

JQSDisabled

JQSEnabled

JQStatus

data JQStatus

Constructors

Creating

Deleted

Deleting

Invalid

Updating

Valid

JobDefinitionType

data JobDefinitionType

Constructors

Container

JobStatus

data JobStatus

Constructors

Failed

Pending

Runnable

Running

Starting

Submitted

Succeeded

AttemptContainerDetail

data AttemptContainerDetail

An object representing the details of a container that is part of a job attempt.

See: attemptContainerDetail smart constructor.

attemptContainerDetail :: AttemptContainerDetail

Creates a value of AttemptContainerDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acdTaskARN - The Amazon Resource Name (ARN) of the Amazon ECS task that is associated with the job attempt.

acdContainerInstanceARN - The Amazon Resource Name (ARN) of the Amazon ECS container instance that hosts the job attempt.

acdReason - A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

acdLogStreamName - Undocumented member.

acdExitCode - The exit code for the job attempt. A non-zero exit code is considered a failure.

acdTaskARN :: Lens' AttemptContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon ECS task that is associated with the job attempt.

acdContainerInstanceARN :: Lens' AttemptContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon ECS container instance that hosts the job attempt.

acdReason :: Lens' AttemptContainerDetail (Maybe Text)

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

acdLogStreamName :: Lens' AttemptContainerDetail (Maybe Text)

Undocumented member.

acdExitCode :: Lens' AttemptContainerDetail (Maybe Int)

The exit code for the job attempt. A non-zero exit code is considered a failure.

AttemptDetail

data AttemptDetail

An object representing a job attempt.

See: attemptDetail smart constructor.

attemptDetail :: AttemptDetail

Creates a value of AttemptDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

adStoppedAt - The Unix timestamp for when the attempt was stopped (when the task transitioned from the RUNNING state to the STOPPED state).

adStartedAt - The Unix timestamp for when the attempt was started (when the task transitioned from the PENDING state to the RUNNING state).

adContainer - Details about the container in this job attempt.

adStatusReason - A short, human-readable string to provide additional details about the current status of the job attempt.

adStoppedAt :: Lens' AttemptDetail (Maybe Integer)

The Unix timestamp for when the attempt was stopped (when the task transitioned from the RUNNING state to the STOPPED state).

adStartedAt :: Lens' AttemptDetail (Maybe Integer)

The Unix timestamp for when the attempt was started (when the task transitioned from the PENDING state to the RUNNING state).

adContainer :: Lens' AttemptDetail (Maybe AttemptContainerDetail)

Details about the container in this job attempt.

adStatusReason :: Lens' AttemptDetail (Maybe Text)

A short, human-readable string to provide additional details about the current status of the job attempt.

ComputeEnvironmentDetail

data ComputeEnvironmentDetail

An object representing an AWS Batch compute environment.

See: computeEnvironmentDetail smart constructor.

computeEnvironmentDetail

Arguments

:: Text
cedComputeEnvironmentName

-> Text
cedComputeEnvironmentARN

-> Text
cedEcsClusterARN

-> ComputeEnvironmentDetail

Creates a value of ComputeEnvironmentDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cedStatus - The current status of the compute environment (for example, CREATING or VALID ).

cedState - The state of the compute environment. The valid values are ENABLED or DISABLED . An ENABLED state indicates that you can register instances with the compute environment and that the associated instances can accept jobs.

cedComputeResources - The compute resources defined for the compute environment.

cedStatusReason - A short, human-readable string to provide additional details about the current status of the compute environment.

cedType - The type of the compute environment.

cedServiceRole - The service role associated with the compute environment that allows AWS Batch to make calls to AWS API operations on your behalf.

cedComputeEnvironmentName - The name of the compute environment.

cedComputeEnvironmentARN - The Amazon Resource Name (ARN) of the compute environment.

cedEcsClusterARN - The Amazon Resource Name (ARN) of the underlying Amazon ECS cluster used by the compute environment.

cedStatus :: Lens' ComputeEnvironmentDetail (Maybe CEStatus)

The current status of the compute environment (for example, CREATING or VALID ).

cedState :: Lens' ComputeEnvironmentDetail (Maybe CEState)

The state of the compute environment. The valid values are ENABLED or DISABLED . An ENABLED state indicates that you can register instances with the compute environment and that the associated instances can accept jobs.

cedComputeResources :: Lens' ComputeEnvironmentDetail (Maybe ComputeResource)

The compute resources defined for the compute environment.

cedStatusReason :: Lens' ComputeEnvironmentDetail (Maybe Text)

A short, human-readable string to provide additional details about the current status of the compute environment.

cedType :: Lens' ComputeEnvironmentDetail (Maybe CEType)

The type of the compute environment.

cedServiceRole :: Lens' ComputeEnvironmentDetail (Maybe Text)

The service role associated with the compute environment that allows AWS Batch to make calls to AWS API operations on your behalf.

cedComputeEnvironmentName :: Lens' ComputeEnvironmentDetail Text

The name of the compute environment.

cedComputeEnvironmentARN :: Lens' ComputeEnvironmentDetail Text

The Amazon Resource Name (ARN) of the compute environment.

cedEcsClusterARN :: Lens' ComputeEnvironmentDetail Text

The Amazon Resource Name (ARN) of the underlying Amazon ECS cluster used by the compute environment.

ComputeEnvironmentOrder

data ComputeEnvironmentOrder

The order in which compute environments are tried for job placement within a queue. Compute environments are tried in ascending order. For example, if two compute environments are associated with a job queue, the compute environment with a lower order integer value is tried for job placement first.

See: computeEnvironmentOrder smart constructor.

computeEnvironmentOrder

Arguments

:: Int
ceoOrder

-> Text
ceoComputeEnvironment

-> ComputeEnvironmentOrder

Creates a value of ComputeEnvironmentOrder with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ceoOrder - The order of the compute environment.

ceoComputeEnvironment - The Amazon Resource Name (ARN) of the compute environment.

ceoOrder :: Lens' ComputeEnvironmentOrder Int

The order of the compute environment.

ceoComputeEnvironment :: Lens' ComputeEnvironmentOrder Text

The Amazon Resource Name (ARN) of the compute environment.

ComputeResource

data ComputeResource

An object representing an AWS Batch compute resource.

See: computeResource smart constructor.

computeResource

Arguments

:: CRType
crType

-> Int
crMinvCPUs

-> Int
crMaxvCPUs

-> Text
crInstanceRole

-> ComputeResource

Creates a value of ComputeResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crEc2KeyPair - The EC2 key pair that is used for instances launched in the compute environment.

crBidPercentage - The minimum percentage that a Spot Instance price must be when compared with the On-Demand price for that instance type before instances are launched. For example, if your bid percentage is 20%, then the Spot price must be below 20% of the current On-Demand price for that EC2 instance.

crSpotIAMFleetRole - The Amazon Resource Name (ARN) of the Amazon EC2 Spot Fleet IAM role applied to a SPOT compute environment.

crImageId - The Amazon Machine Image (AMI) ID used for instances launched in the compute environment.

crDesiredvCPUs - The desired number of EC2 vCPUS in the compute environment.

crTags - Key-value pair tags to be applied to resources that are launched in the compute environment.

crType - The type of compute environment.

crMinvCPUs - The minimum number of EC2 vCPUs that an environment should maintain.

crMaxvCPUs - The maximum number of EC2 vCPUs that an environment can reach.

crInstanceTypes - The instances types that may launched.

crSubnets - The VPC subnets into which the compute resources are launched.

crSecurityGroupIds - The EC2 security group that is associated with instances launched in the compute environment.

crInstanceRole - The Amazon ECS instance role applied to Amazon EC2 instances in a compute environment.

crEc2KeyPair :: Lens' ComputeResource (Maybe Text)

The EC2 key pair that is used for instances launched in the compute environment.

crBidPercentage :: Lens' ComputeResource (Maybe Int)

The minimum percentage that a Spot Instance price must be when compared with the On-Demand price for that instance type before instances are launched. For example, if your bid percentage is 20%, then the Spot price must be below 20% of the current On-Demand price for that EC2 instance.

crSpotIAMFleetRole :: Lens' ComputeResource (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon EC2 Spot Fleet IAM role applied to a SPOT compute environment.

crImageId :: Lens' ComputeResource (Maybe Text)

The Amazon Machine Image (AMI) ID used for instances launched in the compute environment.

crDesiredvCPUs :: Lens' ComputeResource (Maybe Int)

The desired number of EC2 vCPUS in the compute environment.

crTags :: Lens' ComputeResource (HashMap Text Text)

Key-value pair tags to be applied to resources that are launched in the compute environment.

crType :: Lens' ComputeResource CRType

The type of compute environment.

crMinvCPUs :: Lens' ComputeResource Int

The minimum number of EC2 vCPUs that an environment should maintain.

crMaxvCPUs :: Lens' ComputeResource Int

The maximum number of EC2 vCPUs that an environment can reach.

crInstanceTypes :: Lens' ComputeResource [Text]

The instances types that may launched.

crSubnets :: Lens' ComputeResource [Text]

The VPC subnets into which the compute resources are launched.

crSecurityGroupIds :: Lens' ComputeResource [Text]

The EC2 security group that is associated with instances launched in the compute environment.

crInstanceRole :: Lens' ComputeResource Text

The Amazon ECS instance role applied to Amazon EC2 instances in a compute environment.

ComputeResourceUpdate

data ComputeResourceUpdate

An object representing the attributes of a compute environment that can be updated.

See: computeResourceUpdate smart constructor.

computeResourceUpdate :: ComputeResourceUpdate

Creates a value of ComputeResourceUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cruMinvCPUs - The minimum number of EC2 vCPUs that an environment should maintain.

cruMaxvCPUs - The maximum number of EC2 vCPUs that an environment can reach.

cruDesiredvCPUs - The desired number of EC2 vCPUS in the compute environment.

cruMinvCPUs :: Lens' ComputeResourceUpdate (Maybe Int)

The minimum number of EC2 vCPUs that an environment should maintain.

cruMaxvCPUs :: Lens' ComputeResourceUpdate (Maybe Int)

The maximum number of EC2 vCPUs that an environment can reach.

cruDesiredvCPUs :: Lens' ComputeResourceUpdate (Maybe Int)

The desired number of EC2 vCPUS in the compute environment.

ContainerDetail

data ContainerDetail

An object representing the details of a container that is part of a job.

See: containerDetail smart constructor.

containerDetail :: ContainerDetail

Creates a value of ContainerDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdImage - The image used to start the container.

cdCommand - The command that is passed to the container.

cdEnvironment - The environment variables to pass to a container.

cdTaskARN - The Amazon Resource Name (ARN) of the Amazon ECS task that is associated with the container job.

cdUlimits - A list of ulimit values to set in the container.

cdContainerInstanceARN - The Amazon Resource Name (ARN) of the container instance on which the container is running.

cdPrivileged - When this parameter is true, the container is given elevated privileges on the host container instance (similar to the root user).

cdJobRoleARN - The Amazon Resource Name (ARN) associated with the job upon execution.

cdMemory - The number of MiB of memory reserved for the job.

cdUser - The user name to use inside the container.

cdReason - A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

cdLogStreamName - Undocumented member.

cdMountPoints - The mount points for data volumes in your container.

cdExitCode - The exit code to return upon completion.

cdVcpus - The number of VCPUs allocated for the job.

cdReadonlyRootFilesystem - When this parameter is true, the container is given read-only access to its root file system.

cdVolumes - A list of volumes associated with the job.

cdImage :: Lens' ContainerDetail (Maybe Text)

The image used to start the container.

cdCommand :: Lens' ContainerDetail [Text]

The command that is passed to the container.

cdEnvironment :: Lens' ContainerDetail [KeyValuePair]

The environment variables to pass to a container.

cdTaskARN :: Lens' ContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon ECS task that is associated with the container job.

cdUlimits :: Lens' ContainerDetail [Ulimit]

A list of ulimit values to set in the container.

cdContainerInstanceARN :: Lens' ContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) of the container instance on which the container is running.

cdPrivileged :: Lens' ContainerDetail (Maybe Bool)

When this parameter is true, the container is given elevated privileges on the host container instance (similar to the root user).

cdJobRoleARN :: Lens' ContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) associated with the job upon execution.

cdMemory :: Lens' ContainerDetail (Maybe Int)

The number of MiB of memory reserved for the job.

cdUser :: Lens' ContainerDetail (Maybe Text)

The user name to use inside the container.

cdReason :: Lens' ContainerDetail (Maybe Text)

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

cdLogStreamName :: Lens' ContainerDetail (Maybe Text)

Undocumented member.

cdMountPoints :: Lens' ContainerDetail [MountPoint]

The mount points for data volumes in your container.

cdExitCode :: Lens' ContainerDetail (Maybe Int)

The exit code to return upon completion.

cdVcpus :: Lens' ContainerDetail (Maybe Int)

The number of VCPUs allocated for the job.

cdReadonlyRootFilesystem :: Lens' ContainerDetail (Maybe Bool)

When this parameter is true, the container is given read-only access to its root file system.

cdVolumes :: Lens' ContainerDetail [Volume]

A list of volumes associated with the job.

ContainerOverrides

data ContainerOverrides

The overrides that should be sent to a container.

See: containerOverrides smart constructor.

containerOverrides :: ContainerOverrides

Creates a value of ContainerOverrides with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

coCommand - The command to send to the container that overrides the default command from the Docker image or the job definition.

coEnvironment - The environment variables to send to the container. You can add new environment variables, which are added to the container at launch, or you can override the existing environment variables from the Docker image or the job definition.

coMemory - The number of MiB of memory reserved for the job. This value overrides the value set in the job definition.

coVcpus - The number of vCPUs to reserve for the container. This value overrides the value set in the job definition.

coCommand :: Lens' ContainerOverrides [Text]

The command to send to the container that overrides the default command from the Docker image or the job definition.

coEnvironment :: Lens' ContainerOverrides [KeyValuePair]

The environment variables to send to the container. You can add new environment variables, which are added to the container at launch, or you can override the existing environment variables from the Docker image or the job definition.

coMemory :: Lens' ContainerOverrides (Maybe Int)

The number of MiB of memory reserved for the job. This value overrides the value set in the job definition.

coVcpus :: Lens' ContainerOverrides (Maybe Int)

The number of vCPUs to reserve for the container. This value overrides the value set in the job definition.

ContainerProperties

data ContainerProperties

Container properties are used in job definitions to describe the container that is launched as part of a job.

See: containerProperties smart constructor.

containerProperties

Arguments

:: Text
cpImage

-> Int
cpVcpus

-> Int
cpMemory

-> ContainerProperties

Creates a value of ContainerProperties with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpCommand - The command that is passed to the container. This parameter maps to Cmd in the Create a container section of the Docker Remote API and the COMMAND parameter to docker run . For more information, see https://docs.docker.com/engine/reference/builder/#cmd .

cpEnvironment - The environment variables to pass to a container. This parameter maps to Env in the Create a container section of the Docker Remote API and the --env option to docker run . Important: We do not recommend using plain text environment variables for sensitive information, such as credential data.

cpUlimits - A list of ulimits to set in the container. This parameter maps to Ulimits in the Create a container section of the Docker Remote API and the --ulimit option to docker run .

cpPrivileged - When this parameter is true, the container is given elevated privileges on the host container instance (similar to the root user). This parameter maps to Privileged in the Create a container section of the Docker Remote API and the --privileged option to docker run .

cpJobRoleARN - The Amazon Resource Name (ARN) of the IAM role that the container can assume for AWS permissions.

cpUser - The user name to use inside the container. This parameter maps to User in the Create a container section of the Docker Remote API and the --user option to docker run .

cpMountPoints - The mount points for data volumes in your container. This parameter maps to Volumes in the Create a container section of the Docker Remote API and the --volume option to docker run .

cpReadonlyRootFilesystem - When this parameter is true, the container is given read-only access to its root file system. This parameter maps to ReadonlyRootfs in the Create a container section of the Docker Remote API and the --read-only option to docker run .

cpVolumes - A list of data volumes used in a job.

cpImage - The image used to start a container. This string is passed directly to the Docker daemon. Images in the Docker Hub registry are available by default. Other repositories are specified with repository-url /image :tag . Up to 255 letters (uppercase and lowercase), numbers, hyphens, underscores, colons, periods, forward slashes, and number signs are allowed. This parameter maps to Image in the Create a container section of the Docker Remote API and the IMAGE parameter of docker run . * Images in Amazon ECR repositories use the full registry and repository URI (for example, 012345678910.dkr.ecr.region-name.amazonaws.com/repository-name ). * Images in official repositories on Docker Hub use a single name (for example, ubuntu or mongo ). * Images in other repositories on Docker Hub are qualified with an organization name (for example, amazon/amazon-ecs-agent ). * Images in other online repositories are qualified further by a domain name (for example, quay.ioassemblylineubuntu ).

cpVcpus - The number of vCPUs reserved for the container. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run . Each vCPU is equivalent to 1,024 CPU shares.

cpMemory - The hard limit (in MiB) of memory to present to the container. If your container attempts to exceed the memory specified here, the container is killed. This parameter maps to Memory in the Create a container section of the Docker Remote API and the --memory option to docker run .

cpCommand :: Lens' ContainerProperties [Text]

The command that is passed to the container. This parameter maps to Cmd in the Create a container section of the Docker Remote API and the COMMAND parameter to docker run . For more information, see https://docs.docker.com/engine/reference/builder/#cmd .

cpEnvironment :: Lens' ContainerProperties [KeyValuePair]

The environment variables to pass to a container. This parameter maps to Env in the Create a container section of the Docker Remote API and the --env option to docker run . Important: We do not recommend using plain text environment variables for sensitive information, such as credential data.

cpUlimits :: Lens' ContainerProperties [Ulimit]

A list of ulimits to set in the container. This parameter maps to Ulimits in the Create a container section of the Docker Remote API and the --ulimit option to docker run .

cpPrivileged :: Lens' ContainerProperties (Maybe Bool)

When this parameter is true, the container is given elevated privileges on the host container instance (similar to the root user). This parameter maps to Privileged in the Create a container section of the Docker Remote API and the --privileged option to docker run .

cpJobRoleARN :: Lens' ContainerProperties (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that the container can assume for AWS permissions.

cpUser :: Lens' ContainerProperties (Maybe Text)

The user name to use inside the container. This parameter maps to User in the Create a container section of the Docker Remote API and the --user option to docker run .

cpMountPoints :: Lens' ContainerProperties [MountPoint]

The mount points for data volumes in your container. This parameter maps to Volumes in the Create a container section of the Docker Remote API and the --volume option to docker run .

cpReadonlyRootFilesystem :: Lens' ContainerProperties (Maybe Bool)

When this parameter is true, the container is given read-only access to its root file system. This parameter maps to ReadonlyRootfs in the Create a container section of the Docker Remote API and the --read-only option to docker run .

cpVolumes :: Lens' ContainerProperties [Volume]

A list of data volumes used in a job.

cpImage :: Lens' ContainerProperties Text

The image used to start a container. This string is passed directly to the Docker daemon. Images in the Docker Hub registry are available by default. Other repositories are specified with repository-url /image :tag . Up to 255 letters (uppercase and lowercase), numbers, hyphens, underscores, colons, periods, forward slashes, and number signs are allowed. This parameter maps to Image in the Create a container section of the Docker Remote API and the IMAGE parameter of docker run . * Images in Amazon ECR repositories use the full registry and repository URI (for example, 012345678910.dkr.ecr.region-name.amazonaws.com/repository-name ). * Images in official repositories on Docker Hub use a single name (for example, ubuntu or mongo ). * Images in other repositories on Docker Hub are qualified with an organization name (for example, amazon/amazon-ecs-agent ). * Images in other online repositories are qualified further by a domain name (for example, quay.ioassemblylineubuntu ).

cpVcpus :: Lens' ContainerProperties Int

The number of vCPUs reserved for the container. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run . Each vCPU is equivalent to 1,024 CPU shares.

cpMemory :: Lens' ContainerProperties Int

The hard limit (in MiB) of memory to present to the container. If your container attempts to exceed the memory specified here, the container is killed. This parameter maps to Memory in the Create a container section of the Docker Remote API and the --memory option to docker run .

Host

data Host

The contents of the host parameter determine whether your data volume persists on the host container instance and where it is stored. If the host parameter is empty, then the Docker daemon assigns a host path for your data volume, but the data is not guaranteed to persist after the containers associated with it stop running.

See: host smart constructor.

host :: Host

Creates a value of Host with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

hSourcePath - The path on the host container instance that is presented to the container. If this parameter is empty, then the Docker daemon has assigned a host path for you. If the host parameter contains a sourcePath file location, then the data volume persists at the specified location on the host container instance until you delete it manually. If the sourcePath value does not exist on the host container instance, the Docker daemon creates it. If the location does exist, the contents of the source path folder are exported.

hSourcePath :: Lens' Host (Maybe Text)

The path on the host container instance that is presented to the container. If this parameter is empty, then the Docker daemon has assigned a host path for you. If the host parameter contains a sourcePath file location, then the data volume persists at the specified location on the host container instance until you delete it manually. If the sourcePath value does not exist on the host container instance, the Docker daemon creates it. If the location does exist, the contents of the source path folder are exported.

JobDefinition

data JobDefinition

An object representing an AWS Batch job definition.

See: jobDefinition smart constructor.

jobDefinition

Arguments

:: Text
jJobDefinitionName

-> Text
jJobDefinitionARN

-> Int
jRevision

-> Text
jType

-> JobDefinition

Creates a value of JobDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jStatus - The status of the job definition.

jRetryStrategy - The retry strategy to use for failed jobs that are submitted with this job definition.

jParameters - Default parameters or parameter substitution placeholders that are set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

jContainerProperties - An object with various properties specific to container-based jobs.

jJobDefinitionName - The name of the job definition.

jJobDefinitionARN - The Amazon Resource Name (ARN) for the job definition.

jRevision - The revision of the job definition.

jType - The type of job definition.

jStatus :: Lens' JobDefinition (Maybe Text)

The status of the job definition.

jRetryStrategy :: Lens' JobDefinition (Maybe RetryStrategy)

The retry strategy to use for failed jobs that are submitted with this job definition.

jParameters :: Lens' JobDefinition (HashMap Text Text)

Default parameters or parameter substitution placeholders that are set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

jContainerProperties :: Lens' JobDefinition (Maybe ContainerProperties)

An object with various properties specific to container-based jobs.

jJobDefinitionName :: Lens' JobDefinition Text

The name of the job definition.

jJobDefinitionARN :: Lens' JobDefinition Text

The Amazon Resource Name (ARN) for the job definition.

jRevision :: Lens' JobDefinition Int

The revision of the job definition.

jType :: Lens' JobDefinition Text

The type of job definition.

JobDependency

data JobDependency

An object representing an AWS Batch job dependency.

See: jobDependency smart constructor.

jobDependency :: JobDependency

Creates a value of JobDependency with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jJobId - The job ID of the AWS Batch job associated with this dependency.

jJobId :: Lens' JobDependency (Maybe Text)

The job ID of the AWS Batch job associated with this dependency.

JobDetail

data JobDetail

An object representing an AWS Batch job.

See: jobDetail smart constructor.

jobDetail

Arguments

:: Text
jdJobName

-> Text
jdJobId

-> Text
jdJobQueue

-> JobStatus
jdStatus

-> Integer
jdStartedAt

-> Text
jdJobDefinition

-> JobDetail

Creates a value of JobDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jdStoppedAt - The Unix timestamp for when the job was stopped (when the task transitioned from the RUNNING state to the STOPPED state).

jdCreatedAt - The Unix timestamp for when the job was created (when the task entered the PENDING state).

jdRetryStrategy - The retry strategy to use for this job if an attempt fails.

jdAttempts - A list of job attempts associated with this job.

jdDependsOn - A list of job names or IDs on which this job depends.

jdContainer - An object representing the details of the container that is associated with the job.

jdParameters - Additional parameters passed to the job that replace parameter substitution placeholders or override any corresponding parameter defaults from the job definition.

jdStatusReason - A short, human-readable string to provide additional details about the current status of the job.

jdJobName - The name of the job.

jdJobId - The ID for the job.

jdJobQueue - The Amazon Resource Name (ARN) of the job queue with which the job is associated.

jdStatus - The current status for the job.

jdStartedAt - The Unix timestamp for when the job was started (when the task transitioned from the PENDING state to the RUNNING state).

jdJobDefinition - The job definition that is used by this job.

jdStoppedAt :: Lens' JobDetail (Maybe Integer)

The Unix timestamp for when the job was stopped (when the task transitioned from the RUNNING state to the STOPPED state).

jdCreatedAt :: Lens' JobDetail (Maybe Integer)

The Unix timestamp for when the job was created (when the task entered the PENDING state).

jdRetryStrategy :: Lens' JobDetail (Maybe RetryStrategy)

The retry strategy to use for this job if an attempt fails.

jdAttempts :: Lens' JobDetail [AttemptDetail]

A list of job attempts associated with this job.

jdDependsOn :: Lens' JobDetail [JobDependency]

A list of job names or IDs on which this job depends.

jdContainer :: Lens' JobDetail (Maybe ContainerDetail)

An object representing the details of the container that is associated with the job.

jdParameters :: Lens' JobDetail (HashMap Text Text)

Additional parameters passed to the job that replace parameter substitution placeholders or override any corresponding parameter defaults from the job definition.

jdStatusReason :: Lens' JobDetail (Maybe Text)

A short, human-readable string to provide additional details about the current status of the job.

jdJobName :: Lens' JobDetail Text

The name of the job.

jdJobId :: Lens' JobDetail Text

The ID for the job.

jdJobQueue :: Lens' JobDetail Text

The Amazon Resource Name (ARN) of the job queue with which the job is associated.

jdStatus :: Lens' JobDetail JobStatus

The current status for the job.

jdStartedAt :: Lens' JobDetail Integer

The Unix timestamp for when the job was started (when the task transitioned from the PENDING state to the RUNNING state).

jdJobDefinition :: Lens' JobDetail Text

The job definition that is used by this job.

JobQueueDetail

data JobQueueDetail

An object representing the details of an AWS Batch job queue.

See: jobQueueDetail smart constructor.

jobQueueDetail

Arguments

:: Text
jqdJobQueueName

-> Text
jqdJobQueueARN

-> JQState
jqdState

-> Int
jqdPriority

-> JobQueueDetail

Creates a value of JobQueueDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jqdStatus - The status of the job queue (for example, CREATING or VALID ).

jqdStatusReason - A short, human-readable string to provide additional details about the current status of the job queue.

jqdJobQueueName - The name of the job queue.

jqdJobQueueARN - The Amazon Resource Name (ARN) of the job queue.

jqdState - Describes the ability of the queue to accept new jobs.

jqdPriority - The priority of the job queue.

jqdComputeEnvironmentOrder - The compute environments that are attached to the job queue and the order in which job placement is preferred. Compute environments are selected for job placement in ascending order.

jqdStatus :: Lens' JobQueueDetail (Maybe JQStatus)

The status of the job queue (for example, CREATING or VALID ).

jqdStatusReason :: Lens' JobQueueDetail (Maybe Text)

A short, human-readable string to provide additional details about the current status of the job queue.

jqdJobQueueName :: Lens' JobQueueDetail Text

The name of the job queue.

jqdJobQueueARN :: Lens' JobQueueDetail Text

The Amazon Resource Name (ARN) of the job queue.

jqdState :: Lens' JobQueueDetail JQState

Describes the ability of the queue to accept new jobs.

jqdPriority :: Lens' JobQueueDetail Int

The priority of the job queue.

jqdComputeEnvironmentOrder :: Lens' JobQueueDetail [ComputeEnvironmentOrder]

The compute environments that are attached to the job queue and the order in which job placement is preferred. Compute environments are selected for job placement in ascending order.

JobSummary

data JobSummary

An object representing summary details of a job.

See: jobSummary smart constructor.

jobSummary

Arguments

:: Text
jsJobId

-> Text
jsJobName

-> JobSummary

Creates a value of JobSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jsJobId - The ID of the job.

jsJobName - The name of the job.

jsJobId :: Lens' JobSummary Text

The ID of the job.

jsJobName :: Lens' JobSummary Text

The name of the job.

KeyValuePair

data KeyValuePair

A key-value pair object.

See: keyValuePair smart constructor.

keyValuePair :: KeyValuePair

Creates a value of KeyValuePair with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kvpValue - The value of the key value pair. For environment variables, this is the value of the environment variable.

kvpName - The name of the key value pair. For environment variables, this is the name of the environment variable.

kvpValue :: Lens' KeyValuePair (Maybe Text)

The value of the key value pair. For environment variables, this is the value of the environment variable.

kvpName :: Lens' KeyValuePair (Maybe Text)

The name of the key value pair. For environment variables, this is the name of the environment variable.

MountPoint

data MountPoint

Details on a Docker volume mount point that is used in a job's container properties.

See: mountPoint smart constructor.

mountPoint :: MountPoint

Creates a value of MountPoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mpContainerPath - The path on the container at which to mount the host volume.

mpSourceVolume - The name of the volume to mount.

mpReadOnly - If this value is true , the container has read-only access to the volume; otherwise, the container can write to the volume. The default value is false .

mpContainerPath :: Lens' MountPoint (Maybe Text)

The path on the container at which to mount the host volume.

mpSourceVolume :: Lens' MountPoint (Maybe Text)

The name of the volume to mount.

mpReadOnly :: Lens' MountPoint (Maybe Bool)

If this value is true , the container has read-only access to the volume; otherwise, the container can write to the volume. The default value is false .

RetryStrategy

data RetryStrategy

The retry strategy associated with a job.

See: retryStrategy smart constructor.

retryStrategy :: RetryStrategy

Creates a value of RetryStrategy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsAttempts - The number of times to move a job to the RUNNABLE status. You may specify between 1 and 10 attempts. If attempts is greater than one, the job is retried if it fails until it has moved to RUNNABLE that many times.

rsAttempts :: Lens' RetryStrategy (Maybe Int)

The number of times to move a job to the RUNNABLE status. You may specify between 1 and 10 attempts. If attempts is greater than one, the job is retried if it fails until it has moved to RUNNABLE that many times.

Ulimit

data Ulimit

The ulimit settings to pass to the container.

See: ulimit smart constructor.

ulimit

Arguments

:: Int
uHardLimit

-> Text
uName

-> Int
uSoftLimit

-> Ulimit

Creates a value of Ulimit with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uHardLimit - The hard limit for the ulimit type.

uName - The type of the ulimit .

uSoftLimit - The soft limit for the ulimit type.

uHardLimit :: Lens' Ulimit Int

The hard limit for the ulimit type.

uName :: Lens' Ulimit Text

The type of the ulimit .

uSoftLimit :: Lens' Ulimit Int

The soft limit for the ulimit type.

Volume

data Volume

A data volume used in a job's container properties.

See: volume smart constructor.

volume :: Volume

Creates a value of Volume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vName - The name of the volume. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed. This name is referenced in the sourceVolume parameter of container definition mountPoints .

vHost - The contents of the host parameter determine whether your data volume persists on the host container instance and where it is stored. If the host parameter is empty, then the Docker daemon assigns a host path for your data volume, but the data is not guaranteed to persist after the containers associated with it stop running.

vName :: Lens' Volume (Maybe Text)

The name of the volume. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed. This name is referenced in the sourceVolume parameter of container definition mountPoints .

vHost :: Lens' Volume (Maybe Host)

The contents of the host parameter determine whether your data volume persists on the host container instance and where it is stored. If the host parameter is empty, then the Docker daemon assigns a host path for your data volume, but the data is not guaranteed to persist after the containers associated with it stop running.