Network.AWS.Batch.DescribeJobQueues

Description

Describes one or more of your job queues.

Creating a Request

describeJobQueues :: DescribeJobQueues

Creates a value of DescribeJobQueues with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

djqNextToken - The nextToken value returned from a previous paginated DescribeJobQueues request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

djqJobQueues - A list of up to 100 queue names or full queue Amazon Resource Name (ARN) entries.

djqMaxResults - The maximum number of results returned by DescribeJobQueues in paginated output. When this parameter is used, DescribeJobQueues only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeJobQueues request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then DescribeJobQueues returns up to 100 results and a nextToken value if applicable.

data DescribeJobQueues

See: describeJobQueues smart constructor.

Request Lenses

djqNextToken :: Lens' DescribeJobQueues (Maybe Text)

The nextToken value returned from a previous paginated DescribeJobQueues request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

djqJobQueues :: Lens' DescribeJobQueues [Text]

A list of up to 100 queue names or full queue Amazon Resource Name (ARN) entries.

djqMaxResults :: Lens' DescribeJobQueues (Maybe Int)

The maximum number of results returned by DescribeJobQueues in paginated output. When this parameter is used, DescribeJobQueues only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeJobQueues request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then DescribeJobQueues returns up to 100 results and a nextToken value if applicable.

Destructuring the Response

describeJobQueuesResponse

Arguments

:: Int
djqsrsResponseStatus

-> DescribeJobQueuesResponse

Creates a value of DescribeJobQueuesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

djqsrsNextToken - The nextToken value to include in a future DescribeJobQueues request. When the results of a DescribeJobQueues request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

djqsrsJobQueues - The list of job queues.

djqsrsResponseStatus - -- | The response status code.

data DescribeJobQueuesResponse

See: describeJobQueuesResponse smart constructor.

Response Lenses

djqsrsNextToken :: Lens' DescribeJobQueuesResponse (Maybe Text)

The nextToken value to include in a future DescribeJobQueues request. When the results of a DescribeJobQueues request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

djqsrsJobQueues :: Lens' DescribeJobQueuesResponse [JobQueueDetail]

The list of job queues.

djqsrsResponseStatus :: Lens' DescribeJobQueuesResponse Int

- | The response status code.