Network.AWS.Batch.DescribeJobs

Description

Describes a list of AWS Batch jobs.

Creating a Request

describeJobs :: DescribeJobs

Creates a value of DescribeJobs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

djJobs - A space-separated list of up to 100 job IDs.

data DescribeJobs

See: describeJobs smart constructor.

Request Lenses

djJobs :: Lens' DescribeJobs [Text]

A space-separated list of up to 100 job IDs.

Destructuring the Response

describeJobsResponse

Arguments

:: Int
djrsResponseStatus

-> DescribeJobsResponse

Creates a value of DescribeJobsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

djrsJobs - The list of jobs.

djrsResponseStatus - -- | The response status code.

data DescribeJobsResponse

See: describeJobsResponse smart constructor.

Response Lenses

djrsJobs :: Lens' DescribeJobsResponse [JobDetail]

The list of jobs.

djrsResponseStatus :: Lens' DescribeJobsResponse Int

- | The response status code.