Network.AWS.Batch.RegisterJobDefinition

Description

Registers an AWS Batch job definition.

Creating a Request

registerJobDefinition

Arguments

:: Text
rjdJobDefinitionName

-> JobDefinitionType
rjdType

-> RegisterJobDefinition

Creates a value of RegisterJobDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rjdRetryStrategy - The retry strategy to use for failed jobs that are submitted with this job definition. Any retry strategy that is specified during a SubmitJob operation overrides the retry strategy defined here.

rjdParameters - Default parameter substitution placeholders to set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

rjdContainerProperties - An object with various properties specific for container-based jobs. This parameter is required if the type parameter is container .

rjdJobDefinitionName - The name of the job definition to register.

rjdType - The type of job definition.

data RegisterJobDefinition

See: registerJobDefinition smart constructor.

Request Lenses

rjdRetryStrategy :: Lens' RegisterJobDefinition (Maybe RetryStrategy)

The retry strategy to use for failed jobs that are submitted with this job definition. Any retry strategy that is specified during a SubmitJob operation overrides the retry strategy defined here.

rjdParameters :: Lens' RegisterJobDefinition (HashMap Text Text)

Default parameter substitution placeholders to set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

rjdContainerProperties :: Lens' RegisterJobDefinition (Maybe ContainerProperties)

An object with various properties specific for container-based jobs. This parameter is required if the type parameter is container .

rjdJobDefinitionName :: Lens' RegisterJobDefinition Text

The name of the job definition to register.

rjdType :: Lens' RegisterJobDefinition JobDefinitionType

The type of job definition.

Destructuring the Response

registerJobDefinitionResponse

Arguments

:: Int
rjdrsResponseStatus

-> Text
rjdrsJobDefinitionName

-> Text
rjdrsJobDefinitionARN

-> Int
rjdrsRevision

-> RegisterJobDefinitionResponse

Creates a value of RegisterJobDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rjdrsResponseStatus - -- | The response status code.

rjdrsJobDefinitionName - The name of the job definition.

rjdrsJobDefinitionARN - The Amazon Resource Name (ARN) of the job definition.

rjdrsRevision - The revision of the job definition.

data RegisterJobDefinitionResponse

See: registerJobDefinitionResponse smart constructor.

Response Lenses

rjdrsResponseStatus :: Lens' RegisterJobDefinitionResponse Int

- | The response status code.

rjdrsJobDefinitionName :: Lens' RegisterJobDefinitionResponse Text

The name of the job definition.

rjdrsJobDefinitionARN :: Lens' RegisterJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the job definition.

rjdrsRevision :: Lens' RegisterJobDefinitionResponse Int

The revision of the job definition.