Network.AWS.Batch.SubmitJob

Description

Submits an AWS Batch job from a job definition. Parameters specified during SubmitJob override parameters defined in the job definition.

Creating a Request

submitJob

Arguments

:: Text
sjJobName

-> Text
sjJobQueue

-> Text
sjJobDefinition

-> SubmitJob

Creates a value of SubmitJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sjContainerOverrides - A list of container overrides in JSON format that specify the name of a container in the specified job definition and the overrides it should receive. You can override the default command for a container (that is specified in the job definition or the Docker image) with a command override. You can also override existing environment variables (that are specified in the job definition or Docker image) on a container or add new environment variables to it with an environment override.

sjRetryStrategy - The retry strategy to use for failed jobs from this SubmitJob operation. When a retry strategy is specified here, it overrides the retry strategy defined in the job definition.

sjDependsOn - A list of job IDs on which this job depends. A job can depend upon a maximum of 100 jobs.

sjParameters - Additional parameters passed to the job that replace parameter substitution placeholders that are set in the job definition. Parameters are specified as a key and value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

sjJobName - The name of the job. A name must be 1 to 128 characters in length. Pattern: ^[a-zA-Z0-9_]+$

sjJobQueue - The job queue into which the job will be submitted. You can specify either the name or the Amazon Resource Name (ARN) of the queue.

sjJobDefinition - The job definition used by this job. This value can be either a name:revision or the Amazon Resource Name (ARN) for the job definition.

data SubmitJob

See: submitJob smart constructor.

Request Lenses

sjContainerOverrides :: Lens' SubmitJob (Maybe ContainerOverrides)

A list of container overrides in JSON format that specify the name of a container in the specified job definition and the overrides it should receive. You can override the default command for a container (that is specified in the job definition or the Docker image) with a command override. You can also override existing environment variables (that are specified in the job definition or Docker image) on a container or add new environment variables to it with an environment override.

sjRetryStrategy :: Lens' SubmitJob (Maybe RetryStrategy)

The retry strategy to use for failed jobs from this SubmitJob operation. When a retry strategy is specified here, it overrides the retry strategy defined in the job definition.

sjDependsOn :: Lens' SubmitJob [JobDependency]

A list of job IDs on which this job depends. A job can depend upon a maximum of 100 jobs.

sjParameters :: Lens' SubmitJob (HashMap Text Text)

Additional parameters passed to the job that replace parameter substitution placeholders that are set in the job definition. Parameters are specified as a key and value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

sjJobName :: Lens' SubmitJob Text

The name of the job. A name must be 1 to 128 characters in length. Pattern: ^[a-zA-Z0-9_]+$

sjJobQueue :: Lens' SubmitJob Text

The job queue into which the job will be submitted. You can specify either the name or the Amazon Resource Name (ARN) of the queue.

sjJobDefinition :: Lens' SubmitJob Text

The job definition used by this job. This value can be either a name:revision or the Amazon Resource Name (ARN) for the job definition.

Destructuring the Response

submitJobResponse

Arguments

:: Int
sjrsResponseStatus

-> Text
sjrsJobName

-> Text
sjrsJobId

-> SubmitJobResponse

Creates a value of SubmitJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sjrsResponseStatus - -- | The response status code.

sjrsJobName - The name of the job.

sjrsJobId - The unique identifier for the job.

data SubmitJobResponse

See: submitJobResponse smart constructor.

Response Lenses

sjrsResponseStatus :: Lens' SubmitJobResponse Int

- | The response status code.

sjrsJobName :: Lens' SubmitJobResponse Text

The name of the job.

sjrsJobId :: Lens' SubmitJobResponse Text

The unique identifier for the job.