Network.AWS.Batch.UpdateComputeEnvironment

Description

Updates an AWS Batch compute environment.

Creating a Request

updateComputeEnvironment

Arguments

:: Text
uceComputeEnvironment

-> UpdateComputeEnvironment

Creates a value of UpdateComputeEnvironment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uceState - The state of the compute environment. Compute environments in the ENABLED state can accept jobs from a queue and scale in or out automatically based on the workload demand of its associated queues.

uceComputeResources - Details of the compute resources managed by the compute environment. Required for a managed compute environment.

uceServiceRole - The name or full Amazon Resource Name (ARN) of the IAM role that allows AWS Batch to make calls to ECS, Auto Scaling, and EC2 on your behalf.

uceComputeEnvironment - The name or full Amazon Resource Name (ARN) of the compute environment to update.

data UpdateComputeEnvironment

See: updateComputeEnvironment smart constructor.

Request Lenses

uceState :: Lens' UpdateComputeEnvironment (Maybe CEState)

The state of the compute environment. Compute environments in the ENABLED state can accept jobs from a queue and scale in or out automatically based on the workload demand of its associated queues.

uceComputeResources :: Lens' UpdateComputeEnvironment (Maybe ComputeResourceUpdate)

Details of the compute resources managed by the compute environment. Required for a managed compute environment.

uceServiceRole :: Lens' UpdateComputeEnvironment (Maybe Text)

The name or full Amazon Resource Name (ARN) of the IAM role that allows AWS Batch to make calls to ECS, Auto Scaling, and EC2 on your behalf.

uceComputeEnvironment :: Lens' UpdateComputeEnvironment Text

The name or full Amazon Resource Name (ARN) of the compute environment to update.

Destructuring the Response

updateComputeEnvironmentResponse

Arguments

:: Int
ucersResponseStatus

-> UpdateComputeEnvironmentResponse

Creates a value of UpdateComputeEnvironmentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucersComputeEnvironmentName - The name of compute environment.

ucersComputeEnvironmentARN - The Amazon Resource Name (ARN) of the compute environment.

ucersResponseStatus - -- | The response status code.

data UpdateComputeEnvironmentResponse

See: updateComputeEnvironmentResponse smart constructor.

Response Lenses

ucersComputeEnvironmentName :: Lens' UpdateComputeEnvironmentResponse (Maybe Text)

The name of compute environment.

ucersComputeEnvironmentARN :: Lens' UpdateComputeEnvironmentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the compute environment.

ucersResponseStatus :: Lens' UpdateComputeEnvironmentResponse Int

- | The response status code.