Network.AWS.Batch.UpdateJobQueue

Description

Updates a job queue.

Creating a Request

updateJobQueue

Arguments

:: Text
ujqJobQueue

-> UpdateJobQueue

Creates a value of UpdateJobQueue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ujqState - Describes the queue's ability to accept new jobs.

ujqPriority - The priority of the job queue. Job queues with a higher priority (or a lower integer value for the priority parameter) are evaluated first when associated with same compute environment. Priority is determined in ascending order, for example, a job queue with a priority value of 1 is given scheduling preference over a job queue with a priority value of 10 .

ujqComputeEnvironmentOrder - Details the set of compute environments mapped to a job queue and their order relative to each other. This is one of the parameters used by the job scheduler to determine which compute environment should execute a given job.

ujqJobQueue - The name or the Amazon Resource Name (ARN) of the job queue.

data UpdateJobQueue

See: updateJobQueue smart constructor.

Request Lenses

ujqState :: Lens' UpdateJobQueue (Maybe JQState)

Describes the queue's ability to accept new jobs.

ujqPriority :: Lens' UpdateJobQueue (Maybe Int)

The priority of the job queue. Job queues with a higher priority (or a lower integer value for the priority parameter) are evaluated first when associated with same compute environment. Priority is determined in ascending order, for example, a job queue with a priority value of 1 is given scheduling preference over a job queue with a priority value of 10 .

ujqComputeEnvironmentOrder :: Lens' UpdateJobQueue [ComputeEnvironmentOrder]

Details the set of compute environments mapped to a job queue and their order relative to each other. This is one of the parameters used by the job scheduler to determine which compute environment should execute a given job.

ujqJobQueue :: Lens' UpdateJobQueue Text

The name or the Amazon Resource Name (ARN) of the job queue.

Destructuring the Response

updateJobQueueResponse

Arguments

:: Int
ujqrsResponseStatus

-> UpdateJobQueueResponse

Creates a value of UpdateJobQueueResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ujqrsJobQueueARN - The Amazon Resource Name (ARN) of the job queue.

ujqrsJobQueueName - The name of the job queue.

ujqrsResponseStatus - -- | The response status code.

data UpdateJobQueueResponse

See: updateJobQueueResponse smart constructor.

Response Lenses

ujqrsJobQueueARN :: Lens' UpdateJobQueueResponse (Maybe Text)

The Amazon Resource Name (ARN) of the job queue.

ujqrsJobQueueName :: Lens' UpdateJobQueueResponse (Maybe Text)

The name of the job queue.

ujqrsResponseStatus :: Lens' UpdateJobQueueResponse Int

- | The response status code.